Inspect all field values that are pids
Previously, only the `pid` key of the metadata would be inspected before
being passed as the fields value of the map that is encoded by
JSX.encode, but pid can appear in other fields, so all field values are
now escaped with `inspect` if they are pids.  This prevents a MatchError
happening because JSX.encode would return `{:error, :badarg}` if it
tried to encode a pid.